patient can fill symptoms in frontend, bug fixes and ui changes

+++ backend/src/main/java/com/cs490/group4/demo/service/PatientSymptomsService.java
@@ -27,6 +27,7 @@ public class PatientSymptomsService {
 
     @Transactional
     public List<PatientSymptoms> setPatientSymptoms(List<PatientSymptomCreateRequestDTO> symptoms) {
+        if (symptoms.size() == 0) { return List.of(); }
         patientSymptomsRepository.deleteByPatientId(symptoms.get(0).getPatientId());
 
         List<PatientSymptoms> patientSymptoms = symptoms.stream().map(dto -> {
